Use Long type for sample ids in pairs DataFrames, fix id to idxInGroup mapping in general id case. MLTOOLS-5168
ref:2ebd123a341422fac297f8ed48ec2e7eeba53ca1

+++ catboost/spark/catboost4j-spark/core/src/main/scala/ai/catboost/spark/DataSource.scala
@@ -543,8 +543,8 @@ private[spark] object CatBoostPairsDataLoader {
     }
     var schemaWithGroupIdAsStringFields = Seq(
       StructField("groupId", StringType, false),
-      StructField("winnerId", IntegerType, false),
-      StructField("loserId", IntegerType, false)
+      StructField("winnerId", LongType, false),
+      StructField("loserId", LongType, false)
     )
     
     import spark.implicits._
